chore: edit qr, stamp page
