+++ src/pages/Stamp/Component/StampPage/InhaUni.jsx
@@ -5,7 +5,7 @@ import styled from 'styled-components';
 import NavigationBar from '../../../../musterd-ui/NavigationBar';
 
 const TextWrapper = styled.div`
-  width: 100vw;
+  width: 100%;
   height: 143px;
   display: flex;
   flex-direction: column;
